- Clone proxy at root level

+++ tasks/build_proxy.sh
@@ -4,18 +4,13 @@ set -e
 set -v
 
 # Clone Proxy into proxy folder
-git clone https://github.com/EBIBioStudies/BioStudySubmProxy.git proxy
-
-# Go to proxy folder
-cd proxy
-
-ls
+git clone https://github.com/EBIBioStudies/BioStudySubmProxy.git .
 
 # Copy the artifacts generated in build_ui stage
-# if [ -f subtool.tar.gz ]; then
-cp subtool.tar.gz .;
-tar -xvf subtool.tar.gz -C src/main/webapp;
-# fi
+if [ -f subtool.tar.gz ]; then
+  cp subtool.tar.gz .;
+  tar -xvf subtool.tar.gz -C src/main/webapp;
+fi
 
 # Get property values from Environment Variables
 bsBackendUrl=${PROXY_BACKEND_URL}
